+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # CHANGELOG
+## v0.2.0:
+* Add ```data-keys``` options for uses custom data keys.
+## v0.1.2:
+* Add ```exlusive``` options.
+* Add ```event``` and ```data``` options for separate event and data.
 ## v0.1.0:
 * First build.